+++ project/src/draw.h
@@ -5,6 +5,6 @@
 #include "../lib/vbe/vbe.h"
 
 int draw_mouse(hw_mouse_t *mouse, hw_video_t *video);
-int draw_string(hw_video_t *video, const char *str, uint16_t x, uint16_t y, uint32_t bg_color);
+int draw_string(hw_video_t *video, const char *str, int32_t x, int32_t y, uint32_t bg_color);
 
 #endif /* _DRAW_H_ */
